# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Todoタスク管理サービス","description":"HTTP service for todo management","version":"1.0"},"servers":[{"url":"http://localhost:8088"}],"paths":{"/add":{"post":{"tags":["todoservice"],"summary":"addTodoTask todoservice","description":"Todoタスクを追加する。","operationId":"todoservice#addTodoTask","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddTodoTaskRequestBody"},"example":{"asignee":"Voluptatem qui assumenda ipsum.","contents":"Laudantium fugiat iusto eaque placeat officia assumenda.","end_date":"1970-11-01","label":"Ut quia error rerum.","start_date":"2004-04-03","title":"Quo temporibus dolores voluptas doloremque et aut."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"boolean","example":true},"example":false}}}}}},"/todo/{id}":{"delete":{"tags":["todoservice"],"summary":"DeleteTodoTask todoservice","description":"指定したTodoタスクを削除する。","operationId":"todoservice#DeleteTodoTask","parameters":[{"name":"id","in":"path","description":"Todo task id","required":true,"schema":{"type":"integer","description":"Todo task id","example":1515767968551375855,"format":"int64"},"example":8928990878676657841}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"boolean","example":true},"example":true}}}}},"get":{"tags":["todoservice"],"summary":"GetTodoTask todoservice","description":"指定したTodoタスクの詳細を取得する。","operationId":"todoservice#GetTodoTask","parameters":[{"name":"id","in":"path","description":"Todo task id","required":true,"schema":{"type":"integer","description":"Todo task id","example":2174113226990406773,"format":"int64"},"example":4057730328338671830}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoTaskInfo"},"example":{"asignee":"Ut laudantium.","contents":"Dolor eius vel cupiditate magnam saepe mollitia.","end_date":"2009-08-02","id":6101361518690907080,"label":"Ut culpa itaque quae commodi porro maxime.","start_date":"2015-04-02","status":"Aut dolore.","title":"A ipsam delectus qui ducimus enim."}}}}}},"put":{"tags":["todoservice"],"summary":"UpdateTodoTask todoservice","description":"指定したTodoタスクを更新する。","operationId":"todoservice#UpdateTodoTask","parameters":[{"name":"id","in":"path","description":"Todo task id","required":true,"schema":{"type":"integer","description":"Todo task id","example":1432598611792333766,"format":"int64"},"example":2445617283307198598}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"boolean","example":false},"example":false}}}}}},"/todolist":{"get":{"tags":["todoservice"],"summary":"GetTodoTaskList todoservice","description":"Todoタスク一覧を取得する。","operationId":"todoservice#GetTodoTaskList","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TodoTaskTitleList"},"description":"Todo task list","example":[{"id":7879620804179258190,"label":"Velit voluptas labore autem at dolor enim.","status":"Dolores et asperiores.","title":"Aperiam iste sed placeat consequatur doloribus."},{"id":7879620804179258190,"label":"Velit voluptas labore autem at dolor enim.","status":"Dolores et asperiores.","title":"Aperiam iste sed placeat consequatur doloribus."},{"id":7879620804179258190,"label":"Velit voluptas labore autem at dolor enim.","status":"Dolores et asperiores.","title":"Aperiam iste sed placeat consequatur doloribus."},{"id":7879620804179258190,"label":"Velit voluptas labore autem at dolor enim.","status":"Dolores et asperiores.","title":"Aperiam iste sed placeat consequatur doloribus."}]},"example":[{"id":7879620804179258190,"label":"Velit voluptas labore autem at dolor enim.","status":"Dolores et asperiores.","title":"Aperiam iste sed placeat consequatur doloribus."},{"id":7879620804179258190,"label":"Velit voluptas labore autem at dolor enim.","status":"Dolores et asperiores.","title":"Aperiam iste sed placeat consequatur doloribus."}]}}}}}}},"components":{"schemas":{"AddTodoTaskRequestBody":{"type":"object","properties":{"asignee":{"type":"string","description":"タスクを割り当てられた人の名前","example":"Minima impedit nam laudantium non id."},"contents":{"type":"string","description":"Todoタスクの説明","example":"Quia nisi molestias."},"end_date":{"type":"string","description":"Todoタスクの終了予定日","example":"1991-02-15","format":"date"},"label":{"type":"string","description":"Todoタスクのラベル","example":"Ipsam non cumque."},"start_date":{"type":"string","description":"Todoタスクの開始予定日","example":"1994-08-31","format":"date"},"title":{"type":"string","description":"Todoタスクのタイトル","example":"In corrupti laborum sed odio eos."}},"example":{"asignee":"Culpa adipisci.","contents":"Voluptatibus animi et sed facilis.","end_date":"2005-12-18","label":"Asperiores harum excepturi consequatur.","start_date":"1987-03-08","title":"Explicabo mollitia officiis aperiam."},"required":["title","start_date","end_date"]},"TodoTaskInfo":{"type":"object","properties":{"asignee":{"type":"string","description":"タスクを割り当てられた人の名前","example":"Ut explicabo sed nobis provident."},"contents":{"type":"string","description":"Todoタスクの説明","example":"Sit odit minima tenetur."},"end_date":{"type":"string","description":"Todoタスクの終了予定日","example":"1971-12-31","format":"date"},"id":{"type":"integer","description":"TodoタスクのユニークID","example":9192079933244509432,"format":"int64"},"label":{"type":"string","description":"Todoタスクのラベル","example":"Rerum ea et qui."},"start_date":{"type":"string","description":"Todoタスクの開始予定日","example":"2014-02-25","format":"date"},"status":{"type":"string","description":"Todoタスクの進捗状況","example":"Cum voluptatem nam sunt assumenda."},"title":{"type":"string","description":"Todoタスクのタイトル","example":"Voluptatem debitis dolorum vitae est dolores."}},"description":"Todo task","example":{"asignee":"Aliquid labore.","contents":"Quidem veritatis porro impedit tenetur aut deleniti.","end_date":"2013-02-26","id":407848682973057687,"label":"Recusandae est est.","start_date":"1991-12-04","status":"Eos repellat dolorem quibusdam repudiandae praesentium.","title":"Ullam exercitationem provident eveniet doloribus voluptate."},"required":["id","title","status"]},"TodoTaskTitleList":{"type":"object","properties":{"id":{"type":"integer","description":"TodoタスクのユニークID","example":1560018128273690062,"format":"int64"},"label":{"type":"string","description":"Todoタスクのラベル","example":"Voluptatum voluptas rerum qui non."},"status":{"type":"string","description":"Todoタスクの進捗状況","example":"Placeat amet ipsum qui perferendis quia omnis."},"title":{"type":"string","description":"Todoタスクのタイトル","example":"Ut ea ad quia dolores."}},"description":"Todoタスクの情報","example":{"id":2740634361015499265,"label":"Qui soluta qui velit.","status":"Mollitia est veniam autem.","title":"Laboriosam reiciendis optio."},"required":["id","title","status"]}}},"tags":[{"name":"todoservice","description":"Todoタスク管理サービス"}]}